update validator and data component for compiler

+++ xai/formatter/hypertext_markup/writer.py
@@ -590,11 +590,11 @@ class HtmlWriter(Writer):
             table_header = ['Statistical Field', 'Value']
             table_values = list()
             table_values.append(['Total valid count', "%d" % int(num_stats.total_count)])
-            table_values.append(['Min', "%d" % int(num_stats.min)])
-            table_values.append(['Max', "%d" % int(num_stats.max)])
-            table_values.append(['Mean', "%d" % int(num_stats.mean)])
-            table_values.append(['Median', "%d" % int(num_stats.median)])
-            table_values.append(['Standard deviation', "%d" % int(num_stats.sd)])
+            table_values.append(['Min', "%.3f" % float(num_stats.min)])
+            table_values.append(['Max', "%.3f" % float(num_stats.max)])
+            table_values.append(['Mean', "%.3f" % float(num_stats.mean)])
+            table_values.append(['Median', "%.3f" % float(num_stats.median)])
+            table_values.append(['Standard deviation', "%.3f" % float(num_stats.sd)])
             table_values.append(['NAN count', "%d" % int(num_stats.nan_count)])
 
             self.html.article[-1].items.append(self.html.add_table_image_group(
